+++ src/components/cart/SelectedProductItem.tsx
@@ -1,7 +1,6 @@
-import { ChangeEventHandler } from 'react';
 import { css, styled } from 'styled-components';
-import { MAX_NUMBER_LENGTH, NOT_NUMBER, QUANTITY } from '../../constants';
 import { useSetCart } from '../../hooks/useCart';
+import { useHandleQuantityInput } from '../../hooks/useHandleQuantityInput';
 import { useLoadCart } from '../../hooks/useLoadCart';
 import { Product } from '../../types';
 import Button from '../common/Button';
@@ -18,20 +17,12 @@ const SelectedProductItem = ({ id, imageUrl, name, price, quantity }: Props) => 
   const { setIsSelected, setQuantity } = useLoadCart(id);
   const { addToCart, removeItemFromCart } = useSetCart(id);
 
-  const handleNumberInputChange: ChangeEventHandler<HTMLInputElement> = ({ target }) => {
-    const { value } = target;
-
-    if (value === QUANTITY.NONE) {
-      setIsSelected(false);
-      removeItemFromCart();
-
-      return setQuantity(QUANTITY.INITIAL);
-    }
-
-    const onlyTwoDigits = parseInt(value.replace(NOT_NUMBER, '').slice(0, MAX_NUMBER_LENGTH));
-    setQuantity(onlyTwoDigits);
-    addToCart(String(onlyTwoDigits));
-  };
+  const handleNumberInputChange = useHandleQuantityInput({
+    setIsSelected,
+    removeItemFromCart,
+    setQuantity,
+    addToCart,
+  });
 
   return (
     <div>
